allow creating namespaced formatMessage functions
format message configuration is currently uses global values. this
works for most use cases, but there are some cases where it makes
sense to to allow multiple configurations in the same project. this
is especially usefult for libraries that want to use format message
to localize strings without conflicting with other code using format
message.

calling formatMessage.namespace() will return an isolated instance
of formatMessage. calling setup on this instance will not affect any
other instances of format message. the typical use case would be to
have a module that exports the namespaced instance of formatMessage.

importing formatMessage from a different module would break the
static analysis that is used for extracting messages. it now works
with any module named `format-message` even if it is not in
node_modules.