Allow tags in `formatChildren()` messages to have names that are strings.
`formatChildren()` can now accept an Object or Array as the second argument. When it's matching tags to components to substitute, it looks for a key on the `elements` parameter with the matching name. If `elements` is an Array, the integer indexes in the tags function as keys. If `elements` is an object, the tag names are used to find matching properties in the object.

This makes it possible to use names for tags. to provide more context to translators and also to developers. For example:

```
// (React JSX)
formatChildren(formatMessage('There is something <em>important</em> I want you to know. <a>View</a>'), {
  em: <em className="prominent"/>,
  a: <a href="path-to-url"/>
})
```

The tag names may only include characters that are valid in property names (i.e. only upper- and lowercase letters, numbers, underscore, and dollar sign)