Fix 'unique "key" prop' error with react `createChildren()`
Using the react form of `formatChildren()`, the following message appears in the browser console for each React component that's passed to `formatChildren()` that doesn't include a `key` property:

```
Warning: Each child in an array or iterator should have a unique "key" prop. Check the render method of `Component1Indexed`. See https://fb.me/react-warning-keys for more information.
```

This change fixes the problem by dynamically creating a unique key if the `key` property isn't already defined. The key is created as a combination of the element type and a "global" incrementing counter (so multiple elements of the same type don't have duplicate ids).