feat(formatChildren): allow non-numeric tag names
Closes #117

Allow `formatChildren()` to accept an Object or Array as the second argument. The tag names in the message must correspond to properties (or indexes) of the Object or Array. Tag names cannot contain `<`, `/`, `>`, or whitespace. This can provide more context to translators and also to developers. For example:

```
// (React JSX)
formatChildren(formatMessage('There is something <em>important</em> I want you to know. <a>View</a>'), {
  em: <em className="prominent"/>,
  a: <a href="path-to-url"/>
})
```