Make `MODULE_NAME_PATTERN` more flexible (#325)
* Make `MODULE_NAME_PATTERN` more flexible

Right now, format-message looks for itself being imported via the following patterns:

* `import t from 'format-message'`
* `import t from 'foo/format-message'`

In other words, the import must match `'format-message'` exactly or be prepended by a slash. This commit expands upon this to also allow imports that match these patterns:

* `import t from 'foo-format-message'`
* `import t from 'foo.format-message'`

In other words, the character that prepends `format-message` can now be a dash or a period.

This gives format-message consumers more flexibility with naming/importing their own libraries that wrap format-message. (See also: https://github.com/format-message/format-message/issues/232)

* Use a valid Ubuntu pool fur Azure Pipelines

* Update node versions used in Azure Pipelines

* Fix `format-message-interpret` spec

The spec in question incorrectly assumes that a signed zero equals an unsigned zero. Perhaps this comparison used to return true—hence the tests—but it certainly doesn't in my own testing (or in the [Online ICU Message Editor](https://format-message.github.io/icu-message-format-for-translators/editor.html))

* Accept any non-word character prefix in MODULE_NAME_PATTERN

We're prefixing `format-message` with the `\b` RegExp token, which
will match for any character that is not a letter, number, or underscore

* Drop support for Node 10 in Azure Pipelines